backend/requirements.txt
# Core
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0  # PostgreSQL adapter

# Authentication & Security
djangorestframework-simplejwt>=5.3,<6.0
django-otp>=1.3,<2.0  # MFA support
django-two-factor-auth>=1.16,<2.0
argon2-cffi>=23.1,<24.0  # Password hashing

# API
drf-spectacular>=0.27,<1.0  # OpenAPI schema
django-filter>=24.1,<25.0

# Background Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0  # Periodic tasks

# Storage
boto3>=1.34,<2.0  # S3 compatible storage
django-storages>=1.14,<2.0

# Utilities
python-dotenv>=1.0,<2.0
python-dateutil>=2.8,<3.0
pydantic>=2.5,<3.0  # Validation

# Development
django-debug-toolbar>=4.2,<5.0
ipython>=8.18,<9.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# Code Quality
ruff>=0.1,<1.0
mypy>=1.7,<2.0
django-stubs>=4.2,<5.0
